+++ pkg/liquidity-source/uniswap/v2/pool_simulator.go
@@ -95,8 +95,7 @@ func (s *PoolSimulator) CalcAmountOut(param pool.CalcAmountOutParams) (*pool.Cal
 		Fee:            &pool.TokenAmount{Token: s.Info.Tokens[indexIn], Amount: bignumber.ZeroBI},
 		Gas:            defaultGas + extraGasByExchange[s.GetExchange()],
 	}
-	if s.taxHandler.HasSellTax(s.Info.Tokens[indexIn]) ||
-		s.taxHandler.HasBuyTax(s.Info.Tokens[indexOut]) {
+	if effectiveAmountIn != amountIn || amountOut != grossAmountOut {
 		result.SwapInfo = SwapInfo{
 			EffectiveAmountIn: effectiveAmountIn.ToBig(),
 			GrossAmountOut:    grossAmountOut.ToBig(),
