+++ pkg/liquidity-source/uniswap/v2/token-tax/handler.go
@@ -1,11 +1,8 @@
 package tokentax
 
 import (
-	"strings"
-
 	"github.com/holiman/uint256"
 
-	"github.com/KyberNetwork/kyberswap-dex-lib/pkg/entity"
 	"github.com/KyberNetwork/kyberswap-dex-lib/pkg/util/big256"
 )
 
@@ -43,16 +40,4 @@ func deductTax(amount, taxBps *uint256.Int) *uint256.Int {
 	var tax uint256.Int
 	tax.Div(tax.Mul(amount, taxBps), big256.UBasisPoint)
 	return tax.Sub(amount, &tax)
-}
-
-func FindPairedToken(pool entity.Pool, baseTokens map[string]struct{}) string {
-	if len(pool.Tokens) != 2 {
-		return ""
-	}
-	for i, token := range pool.Tokens {
-		if _, ok := baseTokens[strings.ToLower(token.Address)]; ok {
-			return strings.ToLower(pool.Tokens[1-i].Address)
-		}
-	}
-	return ""
 }
